The files were having incorrect true verdict
The property violation occurs when function netif_carrier_off is called from nv_nway_reset and then a model variable ldv_carrier is set to zero. Then the second function register_netdev is called from nv_probe and finally the model variable is checked for non zero. As far as it is not true then the property is violated.