Fix incorrect unsafes verdicts
The Unsafes were found by the model 205_8a which detects 'free without alloc' errors. The model 205_9a, however, does not detect such errors, because ldv_set_remove does nothing in the case if an element was not found in the set. Therfore, the correct verdicts for the tasks are 'safe'.