memory-alloca: remove similar benchmarks
I removed several benchmarks from memory-alloca/ (they are duplicates of
those in termination-memory-alloca/ and also those in memory-unsafe/ so
no loss of benchmarks really occurs).  The removed benchmarks are those
for which I have a very strong confidence are similar to other
benchmarks (which were kept) and test the same feature, not contributing
to the diversity of examples in the category.  This is justified by the
fact that the behaviour (outputs and time) of all tools in SV-COMP'15
was the same.