Fix incorrect unsafe verdict
The Unsafe was found by the model 205_8a which detects 'free without alloc' errors. The model 205_9a, however, does not detect such errors, because ldv_set_remove does nothing in the case if an element was not found in the set. Therefore, the correct verdict for the task is 'safe'.