Add Makefile build system. This will be used to test that the sources files can actually be built by clang and/or gcc.
See ``Makefile-README.md`` for documentation.

The warning flags passed to the compiler are based on the
``check`` script which we can probably remove as the build
system superseeds that script.

Note ``WARNINGS_AS_ERRORS`` is set to off by default because
a lot of the benchmarks would fail to compile otherwise.