Fix incorrect unsafe verdict
The Unsafe was found by the model 205_8a which detects 'free without alloc' errors. The model 205_9a, however, does not detect such errors, because ldv_set_remove does nothing in the case if an element was not found in the set. Therefore, the correct verdict for the task is 'safe'.
The original trace for 205_8a was
  init: LDV_SKBS=0 //set is empty
  can_flush_echo_skb->ldv_skb_free: LDV_SKBS = LDV_SKBS - 1 //success
  alloc_can_err_skb->ldv_skb_alloc: return 0 //fail
  ldv_check_final_state: LDV_SKBS != 0 //free without alloc