Fix undefined behavior of left-shift operation
In bitvector/interleave_bits_true-unreach-call.i, the rhs of a left-shift operation
will be 32 for the last iteration of the loop contained.

It leads to undefined behavior according to C standard.

The proposed fix is based on the original program from which this benchmark was ported.
Its link is https://graphics.stanford.edu/~seander/bithacks.html#InterleaveTableObvious