cleanup duplicate files for termination category
Several files ocurred twice in slightly different versions. One version
was submitted to SV-COMP in the last two years. The other version
was obtained by rewriting the program such that only a restricted syntax
of the C standard is used (see
http://www.termination-portal.org/wiki/C_Integer_Programs for details).
I kept only the files that use the resticted syntax but moved them to
from termination-restricted-stroeder-15 to their original folder.

This modification was discussed with Thomas who submitted the new
versions.