Disable building the ``ldv-linux-3.4-simple`` and ``regression`` directories when building with Clang and CLANG_BUILD_HACK=1 is passed and teach TravisCI to build the benchmarks using Clang.
The ``CLANG_BUILD_HACK`` is a temporary hack so that we can have
TravisCI build the benchmarks using Clang. Future work will hopefully
fix the benchmarks so they can be compiled with Clang and then we
can remove this hack.

By default ``CLANG_BUILD_HACK`` is set to 0 so by default the build
system will still try to build the ``ldv-linux-3.4-simple`` and
``regression`` directories with Clang.