Fix several benchmarks in ``ldv-linux-3.4-simple`` that Clang would not compile due to bad assembly code.
The issue is the implementation of the ffs() and fls() "bitop"
routines. Clang would emit an error like this

```
/home/zakharov/launch/inst/current/envs/linux-3.4/linux-3.4/arch/x86/include/asm/bitops.h:456:23: error: unsupported inline asm: input with type 'long' matching output with type 'int'
     : "rm" (x), "0" (tmp));
                      ^~~
1 error generated.
```

Clang is correct to complain here as the types do not match when the "0"
constraint says that ``tmp`` (has type ``long``) should be stored in the same
location as ``x`` (has type ``int``).

The new implementation of these functions takes the implementation that appears
in the current Linux master branch (when CONFIG_x86_64 is defined). Note that
the ``long tmp = -1;`` line has been commented out rather than removed to avoid
changing line locations.