Refactor the TravisCI build and teach it to build the "preproc" benchmarks in various configurations.
Unfortunately this also required changing how TravisCI is setup
because building the "preproc" benchmarks doesn't really work
on Ubuntu 12.04 LTS (Precise) because the version of CMake is too old.
The versions of CMake available from various PPAs are also too old
2.8.12 is the minimum required version and none of PPAs I could find
provided that or newer.

To fix this TravisCI is now set up to build on Ubuntu 14.04 LTS (Trusty)
but this also required changing how the packages are installed because
TravisCI's ``apt`` add-on only works with Precise. As a consequence we
install packages using ``sudo`` instead which implies that we are using
their legacy infrastructure to build