Use a different approach for detecting the C compiler. Parsing the output of ``$(CC) --version`` is not reliable because some versions of gcc don't show the text "gcc" in the output when invoked as the ``cc`` binary.
Instead we invoke the compiler to preprocess a source file and in
that source file we use macros to output different text depending
on the compiler. This text is then read in ``Makefile.rules`` to
determine the compiler.

To avoid confusion setting ``CC_IS_GCC`` and ``CC_IS_CLANG``
is now done outside of the ``ifndef SKIP_LEVEL`` block so that
running ``make debug_vars`` in the root directory does not give
confusing output.