Fixed undefinedness for bitvector category.
For bitvector/interleave_bits_true-unreach-call.c ,
RV-Match reports undefinedness due to a bitwise left shift
for which the right operand is "equal or bigger" than
the bit size of the left operand. (Line 28).
For us, this is a false positive since GNU C only defines
this operation undefined if the right operand is "bigger"
than the bit size of the left operand.
(https://www.gnu.org/software/gnu-c-manual/gnu-c-manual.html#Bit-Shifting)

In the given case .. << (i + 1), i is always smaller than 32,
so assuming a 32 bit machine model, the operation is defined
due to i+1 <= 32.