Fix a few incompatible library declarations in directory loops.
For n.c24_false-unreach-call.i, this is fixed by doing a fresh
pre-processing, which changes only extern function declarations.

For the other files, this is fixed by manually correcting the definition
of size_t (which is only used in extern function declarations) to be unsigned
or fixing malloc declarations.

Thus the change should not affect the semantics of the files.