Replace benchmarks in loop-industry-pattern with 32bit versions.
These were provided by the original author anushrijana in
https://github.com/sosy-lab/sv-benchmarks/pull/160#issuecomment-261312555.

The changes in these files are more than only a correct preprocessing,
for example there are at least also renamings.
However, as the original files have not really been used yet,
and the new files are from the same author,
I think this is acceptable.

The new files also fix several other problems the compilers detected
as can be seen by the change to the Makefile.