New concurrency benchmarks - Pthread driver races (#151)
* New Concurrency Benchmarks - Pthreaded Drivers

Initial submission of a new group of benchamrks for the sv-comp concurrency
category.

* Update concurrency set files and rename benchmarks

Forgot to add new benchmarks to set files, and didn't have the naming
convention correct for the new benchmarks.  This commit fixes these issues.

* Pthread driver races - include model needed for preprocessing

This commit includes the Model required to close the environment
required by the drivers.

* Switch pthread driver race benchmarks to 32bit architecture

These were originally built and preprocessed with a 64bit
architecture.  This commit switches them to 32bit.

* Update build system for new pthread driver races benchmarks

This was previously overlooked for the initial pull request,
but has now been corrected.

* Update README.txt for pthread driver races to list contributor

These new benchmarks had been attributed to their source, but
the team contributing them to svcomp was not listed.  This commit
corrects this.

* Remove double semi-colon from preprocessed files

These are illegal in struct declarations.  This commit removes
these.

* Removed calls to __VERIFIER_nondet_pthread_t

This function is not defined in the SV-COMP rules, and was also
not needed for these benchmarks. So I completely removed it.

* Fixed a bunch of errors, mainly missing function and structure defs

* Added more missing declarations

* Done with nvram driver

* Done fixing errors and generated preprocessed files

* Removed property file

* Disregard directory Model

* Set C standard to gnu99

* Fixed missing variable initialization

* Fixed an error related to missing argument type

* Taking care of the model folder

* Added symbolic link to license file