Attempt to fix the first benchmark in #191
This benchmark contains dispatch to uninitialized function pointers,
which results in assertion failure. For example, var_group_4 is calloced,
so __cil_tmp23 at line 8285 is equal to the return value of calloc + 1600.
__cil_tmp24 at line 8286 is the deference of __cil_tmp23 which equals to 0 (calloced memory).
 __cil_tmp26 at line 8288 will be 360, which is not initialized.
__cil_tmp27 at line 8289 is the deference of __cil_tmp26 which can be any value.
In line 8295, a function pointer is called which is the deference of __cil_tmp27.
So a function that results in the assertion failure will be called
since *__cil_tmp27 could alias with any function.

For the record, this fix is not supposed to fix all uninitialized memory accesses.
We only intend to fix those that lead to assertion failure.