fixed one file in c/termination-15
The benchmark c/termination-15/array16_alloca_true-termination.c was
incorrectly labeled with a true-termination verdict.

This file consists of two loops that both use an array. The second loop
is terminating if each value of the array is even.
I presume that the first loop should initialize the array with even
numbers. However, it does not. If the value of arr[i] is a negative odd
number, the if condition in the first loop evaluates ot false, because
e.g., -1%2 is -1 in C.

I renamed the original file from ``array16_alloca_true-termination.c``
to ``array16_alloca_original_false-termination.c.i`` and additionally
added the file ``array16_alloca_fixed_true-termination.c`` where I
changed ``arr[i] % 2 == 1`` to ``arr[i] % 2 != 0`` because I presume
that the intension of the original author was to check if the number is
odd.