termination-crafted: fix invalid dereference in NonTermination3
The variable i had arbitrary value when accessing the array a.
I simply swapped the access and getting a new nondet value
(in the first iteration there's a nondet value from the initialization)
so now the loop is like this:

i = __VERIFIER_nondet_int();
while (0 <= i && i < 10 && a[i] >= 0) {
        a[i] = 0;
        i = __VERIFIER_nondet_int();
}

The false-termination still holds, since if i will be always the same, lets say 1,
then the loop will never terminate. The out-of-bound access was removed,
since now the value of i is restricted to [0, 9] by the loop's condition.

Signed-off-by: Marek Chalupa <mchalupa@mail.muni.cz>