busybox: fix bb_errno dereferences
bb_errno is a static pointer, so it points to NULL. However,
it is dereferenced in many of busybox benchmarks. This patch
adds a new varibale bb_errno_location that is initialized to
non-deterministic at the beginning of the main function and
makes bb_errno point to this variable.

The first of the series of patches to fix #385

Signed-off-by: Marek Chalupa <mchalupa@mail.muni.cz>