Define s->version field in ssh client benchmarks
In all of these benchmarks if s->version is not initialised then the bevhaviour is undefined
but includes a path that leads to the ((s->version & 65280) != 768) condition being true
which causes the program to immediately exit. A tool is free to choose this path given
the undefined behaviour. These fixes initialise s->version so that this condition is defined.

In the s3 benchmark some other temporary variables are now initialised to remove some
other warnings.