Replace memcpy with memmove for overlapping memory
Prior to this commit, some benchmarks called memcpy with overlapping
source and destination areas, which is undefined behaviour according to
POSIX.1-2008 and the C standard. The CBMC tool therefore gives a result
of false, including a counterexample, on the benchmarks that are
expected to return true.

These benchmarks, being examples from device drivers, were supposedly
written with overlapping source and memory regions intentionally, and
the code 'works' in practice. However, SV-COMP participants should
arguably not be expected to get the right result for benchmarks relying
on quirky, non-standards-conformant hardware behaviour.

This commit replaces those memcpy calls with calls to memmove, which has
identical semantics but is defined even when called on overlapping
memory regions. Only those calls that CBMC complained about within a
timeout of 15 minutes were removed; thus, safe calls to memcpy, as well
as overlapping calls that CBMC could not discover, remain in the
benchmarks.