Add stubs for all undefined extern functions called by programs in seq-mthreaded
All these functions are void functions and have mostly scalar
parameters (only a few char pointers).
The programs here seem to simulate a controller software,
and the extern functions are there for triggering actions in the
controlled system.
So adding no-op stubs for them should not affect neither the actual nor
the intended semantics.