Add stubs for nonstandard extern functions in SSH programs.
These call only a single function (sslv3_base_method),
which obviously is intended to allocate and return a struct.
Most fields of this struct are function pointers,
but their field names are never referenced,
so we initialize them with null.
There is another unused pointer field,
and a single int field, which gets initialized nondeterministally in the
stub.