ntdrivers memsafety: Remove calls to KeGetCurrentThread and KeSetPriorityThread
These functions are undefined. KeGetCurrentThread is used to get
argument for KeSetPriorityThread. This function should just set
priority of the thread that can change scheduling [1]. But since
these benchmarks are not threaded, it setting the priority behaves like no-op.
Therefore it shoul be safe to just remove the calls to these functions.

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-kesetprioritythread