ntdriver memsafety: comment out calls to undefined MmProbeAndLockPages
This call modifies the attributes of the _MDL struct, but
in one of the changed files the struct is not allocated,
so modifying the arguments would lead to invalid dereference,
and in the other file the struct is allocated, but the referenced
memory is not accessed anywhere as far as I can see.
The rest of the Mm* functions is commented out too.