``IoAllocateErrorLogEntry`` [1]: "The IoAllocateErrorLogEntry routine allocates an error log entry, and returns a pointer to the packet that the caller uses to supply information about an I/O error." "IoAllocateErrorLogEntry returns a pointer to the error log entry, or NULL if a log entry could not be allocated."
Parameters
``IoObject``: Pointer to a device object representing the device on which an I/O error occurred, or to a driver object representing the driver that found an error.
``EntrySize``: Specifies the size, in bytes, of the error log entry to be allocated. This value cannot exceed ``ERROR_LOG_MAXIMUM_SIZE``.

The return value of this function is used in the benchmarks. As the method essentially allocates a chunk of memory of size ``EntrySize`` or returns a null pointer, I propose we stub it with the following.
````
PVOID IoAllocateErrorLogEntry(PVOID IoObject, UCHAR EntrySize){
    void* rtr;
    if(__VERIFIER_nondet_int()){
        rtr = malloc(EntrySize);
    }
    return rtr;
}
````
Note that this requires the removal of a custom (and highly suspicious) malloc function.

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-ioallocateerrorlogentry