``IoAcquireCancelSpinLock`` [1]:
"The IoAcquireCancelSpinLock routine synchronizes cancelable-state transitions for IRPs in a multiprocessor-safe way."

Parameters
``Irql [out]``: Pointer to a variable in which to save the current IRQL for a subsequent call to IoReleaseCancelSpinLock. Usually, the Irql is saved on the stack as a local variable.

The behavior of this function is, similar to ``IoReleaseCancelSpinLock`` (which was already handled in the benchmarks), a NOOP. Therefore, calls to this function are commented out here and the implementation of the function does nothing.

[1] https://msdn.microsoft.com/en-us/library/windows/hardware/ff548196(v=vs.85).aspx