``WRITE_PORT_UCHAR`` [1]:
"The WRITE_PORT_UCHAR routine writes a byte to the specified port address."

Parameters
``Port``: Pointer to the port, which must be a mapped memory range in I/O space.
``Value``: Specifies a byte to be written to the port.

The function writes some value to the hardware ports; the parameters are not changed. Therefore, I propose to model this function as a NOOP.

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ioaccess/nf-ioaccess-write_port_uchar