``PoSetPowerState`` [1]:
"The PoSetPowerState routine notifies the system of a change in the device power state for a device."
"On Windows 2000 and later versions of the operating system, PoSetPowerState returns the previous power state. On Windows 98/Me, PoSetPowerState returns the state passed in State."

Parameters
``DeviceObject``: A pointer to the target DEVICE_OBJECT.
``Type``: Indicates a POWER_STATE_TYPE value. Drivers must specify DevicePowerState.
``State``: Specifies the power state to be set. Drivers must specify a DEVICE_POWER_STATE value.

Since the function returns a state from an external device, returning a nondeterministic value is a good approximation. If, however, the operating system is Win98, the parameter State is returned instead of some other value. I propose to model this in the following way:
```
POWER_STATE PoSetPowerState(PDEVICE_OBJECT DeviceObject ,
                                                                   POWER_STATE_TYPE Type ,
                                                                   POWER_STATE State ) {
   static int isFixed = 0;
   static int isWin98 = 0;
   if (isFixed == 0)
   {
	   isWin98 = __VERIFIER_nondet_int();
	   isFixed = 1;
   }

   if (isWin98)
   {
       return State;
   }

   int value = __VERIFIER_nondet_int();
   __VERIFIER_assume(value >= 0 && value <= 7);

   POWER_STATE pwr;
   pwr.SystemState = value;
   return pwr;
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ntifs/nf-ntifs-posetpowerstate