``PoRequestPowerIrp`` [1]:
"The PoRequestPowerIrp routine allocates a power IRP and sends it to the top driver in the device stack for the specified device."
"PoRequestPowerIrp returns one of the following:
   Return code                     Description
   STATUS_PENDING                  The IRP has been sent.
   STATUS_INSUFFICIENT_RESOURCES   The routine could not allocate the IRP.
   STATUS_INVALID_PARAMETER_2      MinorFunction does not signify a valid minor power IRP code."

Parameters
``DeviceObject``: A pointer to the target DEVICE_OBJECT for the IRP. In Windows 2000 and later versions of Windows, this parameter can point to a physical device object (PDO) or a functional device object (FDO). In Windows 98/Me, this parameter must point to the PDO of the underlying device.
``MinorFunction``: Specifies one of the following minor power IRP codes: IRP_MN_QUERY_POWER, IRP_MN_SET_POWER, or IRP_MN_WAIT_WAKE.
``PowerState``: Specifies a POWER_STATE value to pass in the IRP. For IRP_MN_SET_POWER and IRP_MN_QUERY_POWER, specify the requested new device power state. Possible values are DEVICE_POWER_STATE values. For IRP_MN_WAIT_WAKE, specify the lowest (least-powered) system power state from which the device should be allowed to wake the system. Possible values are SYSTEM_POWER_STATE values.
``CompletionFunction``: A pointer to the caller's PowerCompletion callback routine. The I/O manager calls this routine when the IRP has completed. This parameter is optional and can be set to NULL if no PowerCompletion callback routine is needed.
``Context``: A pointer to a caller-supplied context to be passed through to the PowerCompletion callback. When the caller requests a device set-power IRP in response to a system set-power IRP, Context should contain the system set-power IRP that triggered the request.
``Irp``: A pointer to a caller-supplied variable in which PoRequestPowerIrp returns a pointer to the IRP that it allocates. Specify a value for Irp only if the MinorFunction parameter is set to IRP_MN_WAIT_WAKE. Otherwise, this parameter should always be NULL because the IRP might be completed before PoRequestPowerIrp returns, causing this parameter to point to memory that has already been discarded.

As per the definition in [1], the function may return only one of three values. The values are defined as [2]:
STATUS_PENDING                  0x00000103
STATUS_INSUFFICIENT_RESOURCES   0xC000009A
STATUS_INVALID_PARAMETER_2      0xC00000F0

Therefore, I propose the following implementation:
```
NTSTATUS PoRequestPowerIrp(PDEVICE_OBJECT DeviceObject ,
                           UCHAR MinorFunction ,
                           POWER_STATE PowerState ,
                           void (*CompletionFunction)(PDEVICE_OBJECT DeviceObject ,
                                                      UCHAR MinorFunction ,
                                                      POWER_STATE PowerState ,
                                                      PVOID Context ,
                                                      PIO_STATUS_BLOCK IoStatus ) ,
                                                      PVOID Context ,
                                                      PIRP *Irp ) {
	NTSTATUS rtr = __VERIFIER_nondet_long();
	__VERIFIER_assume(rtr == 0x00000103 /* STATUS_PENDING */ ||
			  rtr == 0xC000009A /* STATUS_INSUFFICIENT_RESOURCES */ ||
			  rtr == 0xC00000F0 /* STATUS_INVALID_PARAMETER_2 */);

	return rtr;
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-porequestpowerirp
[2] https://msdn.microsoft.com/en-us/library/cc704588.aspx