``KeSynchronizeExecution`` [1]:
"The KeSynchronizeExecution routine synchronizes the execution of the specified routine with the interrupt service routine (ISR) that is assigned to a set of one or more interrupt objects."
"KeSynchronizeExecution returns TRUE if the operation succeeds. Otherwise, it returns FALSE."

Parameters
``Interrupt``: A pointer to a set of interrupt objects. The caller obtained this pointer from the IoConnectInterrupt or IoConnectInterruptEx routine.
``SynchronizeRoutine``: Specifies a caller-supplied SynchCritSection routine whose execution is to be synchronized with the execution of the ISR assigned to the interrupt objects.
``SynchronizeContext``: A pointer to a caller-supplied context value to be passed to the SynchCritSection routine when it is called.

This function either returns true or false which is modeled in the following way:
```
BOOLEAN KeSynchronizeExecution(PKINTERRUPT Interrupt ,
                               BOOLEAN (*SynchronizeRoutine)(PVOID SynchronizeContext ) ,
                               PVOID SynchronizeContext ) {
	return __VERIFIER_nondet_int();
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-kesynchronizeexecution