``KeQueryTickCount`` [1]:
"The KeQueryTickCount routine maintains a count of the interval timer interrupts that have occurred since the system was booted."

Parameters
``CurrentCount``: Pointer to the tick count value on return from KeQueryTickCount.

This function writes the TickCount value to the pointer ``CurrentCount``.

I propose to model this function as follows:
```
void KeQueryTickCount(LARGE_INTEGER *i) {
	(*i).u.LowPart = __VERIFIER_nondet_ulong();
	(*i).u.HighPart = __VERIFIER_nondet_long();
}

```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-kequerytickcount