``KeInsertQueueDpc`` [1]:
"The KeInsertQueueDpc routine queues a DPC for execution."
"If the specified DPC object is not currently in a DPC queue, KeInsertQueueDpc queues the DPC and returns TRUE."

Parameters
``Dpc``: Pointer to the KDPC structure for the DPC object. This structure must have been initialized by either KeInitializeDpc or KeInitializeThreadedDpc.
``SystemArgument1``: Specifies driver-determined context data. This value is passed as the SystemArgument1 parameter to the DPC object's CustomDpc or CustomThreadedDpc routine.
``SystemArgument2``: Specifies driver-determined context data. This value is passed as the SystemArgument2 parameter to the DPC object's CustomDpc or CustomThreadedDpc routine.

Proposed model:
```
BOOLEAN KeInsertQueueDpc(PRKDPC Dpc , PVOID SystemArgument1 ,
                         PVOID SystemArgument2 ) {
	return __VERIFIER_nondet_int();
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-keinsertqueuedpc