``KeInitializeDpc`` [1]:
"The KeInitializeDpc routine initializes a DPC object, and registers a CustomDpc routine for that object."

Parameters
``Dpc``: Pointer to a KDPC structure that represents the DPC object to initialize. The caller must allocate storage for the structure from resident memory.
``DeferredRoutine``: Pointer to the CustomDpc routine to associate with the DPC.
``DeferredContext``: Specifies the value to pass as the DeferredContext parameter to CustomDpc.

I propose to initialize the Dpc object with the information provided by the function, while leaving other values undetermined:
```
void KeInitializeDpc(PRKDPC Dpc , void (*DeferredRoutine)(struct _KDPC *Dpc ,
                                                          PVOID DeferredContext ,
                                                          PVOID SystemArgument1 ,
                                                          PVOID SystemArgument2 ) ,
                     PVOID DeferredContext ) {
	Dpc->Type = __VERIFIER_nondet_short();
	Dpc->Number = __VERIFIER_nondet_char();
	Dpc->Importance = __VERIFIER_nondet_char();
	Dpc->DeferredRoutine = DeferredRoutine;
	Dpc->DeferredContext = DeferredContext;
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-keinitializedpc