``ObfReferenceObject`` [1]:
"The ObReferenceObject routine increments the reference count to the given object."
"ObReferenceObject returns a value that is reserved for system use. Drivers must treat this value as VOID."

Parameters
``Object``: Pointer to the object. The caller obtained this parameter either when it created the object or from a preceding call to ObReferenceObjectByHandle after it opened the object.

"ObReferenceObject simply increments the pointer reference count for an object, without making any access checks on the given object, as ObReferenceObjectByHandle and ObReferenceObjectByPointer do."

Therefore, I propose a NOOP:
```
void ObfReferenceObject(PVOID Object ) {
        return;
}
```

[1] https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/wdm/nf-wdm-obfreferenceobject