Remove the make files for the Java tasks.
The information in there (how to compile the tasks)
is not required, as it is defined in the documentation.
Automatic checks for compilation can be done using BenchExec
and compile.xml, which has the advantage that it also immediately
checks that the information in the YAML files is correct
(no chance of mismatching information between Makefiles and YAML files).

Because using compile.xml for checking compilation
will only be possible when BenchExec supports YAML files
and a tool-info module for javac is added,
the checks are now removed from CI,
but will be re-added later.