Replace memcpy with memmove for overlapping memory
This commit replaces calls to `memcpy` with calls to `memmove` where the
source and destination addresses could potentially overlap, and adds an
extern declaration of `memmove` to those files if they did not already
have one. The overlaps were detected using the CBMC Path tool. Calling
`memcpy` on such overlapping memory areas is undefined behaviour, so
CBMC returns a counterexample on benchmarks that are expected to return
true.

This commit fixes the same issues that 55d4083f does for more source
files, but also correctly includes extern declarations of memmove as
introduced in the bugfix 1280ba1f.