Correcting expected verdict for UnsatAddition02
We think that that the assertion is not reachable. The method "addition" calculates the sum of the parameters, i.e. result=m+n. Then the only condition to reach the assertion is: m>=100&n>=100&result<200, which is *unsat* given that result=m+n. We want to ask for your opinion.