Exclude Systems_DeviceDriversLinux64_ReachSafety* from preprocessing-consistency check
Checking these categories takes too much time (over an hour) and CI always fails.
This was not a problem previously because #include statements
got added to these tasks just recently.

Checking preprocessing consistency is not so important for these tasks
as for other categories, because only files with common definitions are
included, no system headers (so we an alway regenerate the preprocessed
files without difficulties).