Fix typing of kfree in ldv-linux-3.4-simple
The .c files did not match the preprocessed files, and the model only matched
(some of!) the preprocessed files. Now uses the same parameter type as in all
other folders. If the intention was to not have qualifiers such as "const" in
the -simple folder: this must previously have been achieved via text
replacement, but it was done in an incomplete way (only the first occurrence of
"const" on each line was removed). So it seems fine to re-introduce some uses of
"const".