Fix source files to ensure re-preprocessing works
Changes original source files to match whatever is now in the
preprocessed files, which sometimes have been updated without making
matching changes to the original source. Most of these changes are
whitespace-only or an order of declarations. Some changes do affect
behaviour, but the competition always uses preprocessed files and thus
this these changes do not affect the intended verdict.