Introduce yml task-definition files
Each benchmark task is now represented by
a task definition in YAML format that describes
all possible benchmark properties and their expected verdicts (if
any).

Since it is obsolete now, we remove the properties and verdicts
from the task's file names. If a task is a preprocessed C file
and the original file exists next to it, it is also renamed.
Since same file names may confuse benchexec,
append a suffix "-$COUNT" to tasks with the same new name (when they
are in different directories)

The script used to do all of this can be found here:
https://gist.github.com/lembergerth/4d98294c51875a9ba85427d13f2c72ea