Adjust compare.sh blacklist and allow wildcards.
With task definition files, the .set files include more
tasks through wildcards than before, and the compare.sh script
does not check for properties. Thus, it is not the right
way to exclude certain sets anymore, but the tasks
themselves should be specified.
To make this specification easier for large number of files,
allow wildcards in the blacklist definition.