Provide script that allows to get all tasks with a certain property
In case people do not use the benchexec infrastructure,
it is difficult to get the full set of benchmark files
with a certain property.

This script makes that task easier:
It gets a property name and, optionally, the benchmark directory,
and prints all tasks with the given property to stdout.
At the moment, properties must be given without a directory prefix to
reliably work. This way, it is not necessary to look at each single task
definition file separately, but it may be imprecise in case
two different properties with the same basename exist in different
directories.

Example usage:
```
./get_tasks_for_property.sh coverage-branches.prp c/
```