Adding a property that should fail in the Wheel Brake System as a benchmark
This version of Wheel Brake System has been part of the SPF repository
for the past few years. It can be found at https://github.com/SymbolicPathFinder/jpf-symbc/blob/master/src/examples/WBS.java. In this commit, I am dding it to the SV-COMP repository of benchmarks so that other tools can also benefit from it.