Adding a equivalence-check on the Replace benchmark
The replace program is part of the Siemens suite of programs. This
benchmark performs pattern matching and substitution. One simple
property that should fail on this benchmark is that the outputs of the
benchmark from running it twice should not be equivalent if the starting
states of the benchmark were not equivalent. This is done by not
resetting the benchmark to its starting state. This benchmark was
obtained from Wang et al.
(https://labs.xjtudlc.com/labs/wlaq/hjwang/toolbench.html).