loops/s3.i: Replace use of nondet_pointer by nondet_ulong
The semantics of __VERIFIER_nondet_pointer, particular whether it does or does
not perform memory allocation, have never been clarified (see issue #767). To
avoid any such ambiguity, just cast a non-deterministic unsigned long to a
pointer. As the benchmark just performs equality tests on the pointer value and
never attempts to dereference, this is a sufficient replacement.