aws-c-common: Replace all uses of nondet_pointer by nondet_ulong
The semantics of __VERIFIER_nondet_pointer, particular whether it does
or does not perform memory allocation, have never been clarified (see
issue #767). To avoid any such ambiguity, just cast a non-deterministic
unsigned long to a void* instead, making clear that no memory allocation
is performed. Any dereferencing of the resulting pointer could be
undefined behaviour, which would expose a bug as the only use of
non-deterministic pointers is to model the behaviour of realloc. This
error case is explicitly called out in comments in the source code.

The changes were applied as follows:
```
perl -p -i -e \
  's/return __VERIFIER_nondet_pointer\(\);/return (void *)__VERIFIER_nondet_ulong();/' \
  c/aws-c-common/*.[chi]
sed -i \
  '/^extern void \*__VERIFIER_nondet_pointer();/d' \
  c/aws-c-common/*.[chi]
```