ldv-linux-3.14-races: Fix original source to match preprocessed version
In 99652449be3, an initialisation fix was added, but the fix was only
applied to the preprocessed code. The original source was left
unmodified. This change makes sure that re-preprocessing does not result
in reverting the fix.

CI did not catch this problem, because
Systems_DeviceDriversLinux64_ReachSafety is exempted for performance
reasons.