Specialise kmem_cache_zalloc to the desired allocation size
kmem_cache_zalloc should allocate (from a cache) memory of the size
confirmed when the cache was created (via kmem_cache_create). Since none
of the benchmarks provide a definition of kmem_cache_create (or struct
kmem_cache), this information cannot be communicated. Instead, hard-code
the allocation size as used in the specific device driver in
kmem_cache_zalloc and use ldv_zalloc instead of ldv_zalloc_unknown_size.
This is in preparation of removing uses of __VERIFIER_nondet_pointer
(see #767).