Remove unused ldv_zalloc_unknown_size
Removes one of the uses of external_alloc, which is implemented using
__VERIFIER_nondet_pointer. As ldv_zalloc_unknown_size is never actually
called, removing the definition is sufficient to get rid of this
particular use of __VERIFIER_nondet_pointer. This is in preparation of
removing uses of __VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
perl -i -e \
  '$p = 1;
   while(<>) {
     if(/^void \*ldv_zalloc_unknown_size\(void\)/) {
       if(/;$/) { $p = 1; } else { $p = 2; }
     }
     elsif($p == 2 || $p == 3) { if(/^\}$/) { $p++; } }
     else { print; }
   }' \
  c/ldv-*/*.[ci]
```