Implement ldv_vzalloc_{,node_}* using ldv_zalloc
Removes one of the uses of ldv_malloc_unknown_size, which is implemented
using __VERIFIER_nondet_pointer. The size is known, thus using
ldv_zalloc is perfectly safe, and also more consistent as it will ensure
zero-initialisation. This is in preparation of removing uses of
__VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
for f in c/ldv-*/*.[ci]
do
perl -i -e \
  '$p = 1; $d = 0;
   while(<>) {
     if(/^void \*ldv_zalloc\(size_t size \) ;$/) {
       next if($d == 2);
       $d = 1;
     }
     if(/^static void \*ldv_vzalloc_(node_)?\d+(___\d+)?\(unsigned long ldv_func_arg1( , int ldv_func_arg2)? \) ?$/) {
       $p = 2;
       if($d == 0) {
         print "void *ldv_zalloc(size_t size ) ;\n";
         $d = 2;
         print;
         next;
       }
     }
     if($p == 2) {
       if(/^  tmp = ldv_malloc_unknown_size\(\);/) {
         print "  tmp = ldv_zalloc(ldv_func_arg1);\n";
         $p = 1;
         next;
       }
       elsif(/^\}$/) { $p = 1; }
     }
     print;
   }' \
  $f
done
```