Implement ldv_dma_pool_alloc_* using ldv_malloc
Removes one of the uses of ldv_malloc_unknown_size, which is implemented
using __VERIFIER_nondet_pointer. The size is inferred as described
below.  This is in preparation of removing uses of
__VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
for f in c/ldv-*/*.[ci]
do
perl -i -e \
  '$p = 1; $d = 0;
   while(<>) {
     if(/^void \*ldv_malloc\(size_t size \) ;$/) {
       next if($d == 2);
       $d = 1;
     }
     if(/^static void \*ldv_dma_pool_alloc_\d+\(struct dma_pool \*ldv_func_arg1 , gfp_t flags , ?$/) {
       $p = 2;
       if($d == 0) {
         print "void *ldv_malloc(size_t size ) ;\n";
         $d = 2;
       }
     }
     if($p == 2) {
       if(/^  tmp = ldv_malloc_unknown_size\(\);/) {
         print "  tmp = ldv_malloc(FIXME);\n";
         $p = 1;
         next;
       }
       elsif(/^\}$/) { $p = 1; }
     }
     print;
   }' \
  $f
done
```
The actual allocation size (the "FIXME" in the above) was then inserted
manually by looking at the `dma_pool_create` calls. This is necessary as
the definition of `struct dma_pool` is not included in the benchmarks,
which would hold the allocation size as one of its members. In two
cases, the size also is an over-approximation as two different pools are
in use and either 4096 or 256 bytes could be required (thus always using
4096 bytes).